updated landing page and created nav
