+++ src/components/LandingPage/LandingPage.css
@@ -7,53 +7,29 @@ body, html {
   font-family: 'Inter', sans-serif;
 }
 
-.nav-button {
-  font-family: 'Inter', sans-serif;
-  font-weight: bold; 
-  background: none;
-  border: none;
-  color:white;
-  font-size: 20px;
-  cursor: pointer;
-}
-
-.container {
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.landing-header {
-  padding: 20px;
-}
-
-.nav-menu {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  justify-content: center;
-  gap: 200px;
-  white-space: nowrap;
-}
-
 .landing-main {
   flex-grow: 1;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  text-align: center;
+  justify-content: space-between;
 }
 
 .main-subtitle {
-  color: white;
-  font-size: 20px;
+  color: #ebe5e5;
+  font-size: 35px;
 }
 
 .main-logo {
-  width: 900px;
+  width: 950px;
   max-width: 100%;
   height: auto;
-  margin-bottom: 20px;
+  margin-top: 80px;
+}
+
+.instaicon {
+  font-size: 50px !important;
+  color: black !important;
 }
