+++ src/components/Nav/Nav.css
@@ -1,42 +1,46 @@
-.nav {
-  /* flex box layout */
+.nav-button {
+    border: 3px solid black;
+    background: #ebe5e5;
+    padding: 15px;
+    display: inline-block;
+    color: black;
+    font-size: 25px;
+}
+
+.nav-menu {
+  list-style-type: none;
+  padding: 0;
+  margin: 22px;
+  display: flex;
+  white-space: nowrap;
+  padding: 0;
+  justify-content: center;
+  gap: 100px;
+  list-style-type: none;
+  flex-grow: 1;
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+}
+
+.nav-menu li {
+  margin: 0 5px; /* Adjust spacing between buttons */
+}
+
+.nav-logo {
+  width: 150px;
+  height: auto;
+  margin-left: 20px;
+  justify-content: left;
+
+}
+
+.nav-container {
   display: flex;
   justify-content: space-between;
   align-items: center;
-
-  margin: 0 0 30px;
-  background-color: #00acb0;
-  overflow: hidden;
+  padding-top: 18px;
 }
 
-.nav a {
-  text-decoration: none;
-}
 
-.nav-title {
-  font-size: 24px;
-  font-weight: 700;
-  color: #f2f2f2;
-  padding-left: 10px;
-  margin: 0;
-}
 
-.navLink {
-  display: inline-block;
-  color: #f2f2f2;
-  background-color: #00acb0;
-  text-align: center;
-  padding: 24px 10px;
-  text-decoration: none;
-  font-size: 15px;
-
-  /* these last two make the logout button in
-  the nav bar look like another link */
-  border: none;
-  cursor: pointer;
-  outline: 0;
-}
-
-.navLink:hover {
-  background-color: #008183;
-}
